Refactored Inventory into Map Inventory Pt1
Should still move the count and update functions OUT of inventory.
