+++ Map_Inventory.cpp
@@ -2,7 +2,7 @@
 
 #include <BWAPI.h>
 #include "Source\CUNYAIModule.h"
-#include "Source\InventoryManager.h"
+#include "Source\Map_Inventory.h"
 #include "Source\Unit_Inventory.h"
 #include "Source\Resource_Inventory.h"
 #include <algorithm>
