+++ Resource_Inventory.cpp
@@ -4,7 +4,7 @@
 #include "Source\CUNYAIModule.h"
 #include "Source\Resource_Inventory.h"
 #include "Source\Unit_Inventory.h"
-#include "Source\InventoryManager.h"
+#include "Source\Map_Inventory.h"
 
 //Resource_Inventory functions.
 //Creates an instance of the resource inventory class.
